Move expression conversion into its own module
